cleanup optimizer, tests
