+++ crates/polars-plan/src/logical_plan/optimizer/count_star.rs
@@ -1,3 +1,5 @@
+use std::path::PathBuf;
+
 use super::*;
 
 pub(super) struct CountStar;
@@ -8,21 +10,18 @@ impl CountStar {
     }
 }
 
-impl OptimizationRule for CountStar {
-    fn optimize_plan(
-        &mut self,
-        lp_arena: &mut Arena<ALogicalPlan>,
-        _expr_arena: &mut Arena<AExpr>,
+impl CountStar {
+    fn is_simple_count_star_plan(
         node: Node,
-    ) -> Option<ALogicalPlan> {
-        // Replace select count(*) from datasource with fast pass map function
-
-        // Do not allow nested structures. Logical plan should be Project -> Scan
+        lp_arena: &mut Arena<ALogicalPlan>,
+    ) -> Option<(Arc<[PathBuf]>, FileScan)> {
+        // Top node should be a projection
         let ALogicalPlan::Projection { input, .. } = lp_arena.get(node) else {
             return None;
         };
 
-        let ALogicalPlan::Scan {
+        // Leaf node should be a scan without predicates
+        let scan @ ALogicalPlan::Scan {
             paths,
             scan_type,
             predicate: None,
@@ -32,27 +31,39 @@ impl OptimizationRule for CountStar {
             return None;
         };
 
+        // Do not support anonymous scans
         if matches!(scan_type, FileScan::Anonymous { .. }) {
             return None;
         }
 
-        let paths = paths.clone();
-        let scan_type = scan_type.clone();
+        Some((paths.clone(), scan_type.clone()))
+    }
+}
 
-        let placeholder = ALogicalPlan::DataFrameScan {
-            df: Arc::new(Default::default()),
-            schema: Arc::new(Default::default()),
-            output_schema: None,
-            projection: None,
-            selection: None,
-        };
-        let placeholder_node = lp_arena.add(placeholder);
+impl OptimizationRule for CountStar {
+    // Replace select count(*) from datasource with specialized map function
+    fn optimize_plan(
+        &mut self,
+        lp_arena: &mut Arena<ALogicalPlan>,
+        _expr_arena: &mut Arena<AExpr>,
+        node: Node,
+    ) -> Option<ALogicalPlan> {
+        CountStar::is_simple_count_star_plan(node, lp_arena).map(|(paths, scan_type)| {
+            let placeholder = ALogicalPlan::DataFrameScan {
+                df: Arc::new(Default::default()),
+                schema: Arc::new(Default::default()),
+                output_schema: None,
+                projection: None,
+                selection: None,
+            };
+            let placeholder_node = lp_arena.add(placeholder);
 
-        let alp = ALogicalPlan::MapFunction {
-            input: placeholder_node,
-            function: FunctionNode::Count { paths, scan_type },
-        };
-        lp_arena.replace(node, alp.clone());
-        Some(alp)
+            let alp = ALogicalPlan::MapFunction {
+                input: placeholder_node,
+                function: FunctionNode::Count { paths, scan_type },
+            };
+            lp_arena.replace(node, alp.clone());
+            alp
+        })
     }
 }
